fix(hog): 0-indexed array errors to resolver
