+++ posthog/hogql/resolver.py
@@ -691,6 +691,9 @@ class Resolver(CloningVisitor):
     def visit_array_access(self, node: ast.ArrayAccess):
         node = super().visit_array_access(node)
 
+        if self.dialect == "clickhouse" and isinstance(node.property, ast.Constant) and node.property.value == 0:
+            raise QueryError("SQL indexes start from one, not from zero. E.g: array[1]")
+
         array = node.array
         while isinstance(array, ast.Alias):
             array = array.expr
@@ -718,6 +721,9 @@ class Resolver(CloningVisitor):
 
     def visit_tuple_access(self, node: ast.TupleAccess):
         node = super().visit_tuple_access(node)
+
+        if self.dialect == "clickhouse" and node.index == 0:
+            raise QueryError("SQL indexes start from one, not from zero. E.g: array.1")
 
         tuple = node.tuple
         while isinstance(tuple, ast.Alias):
